// SGUEES-SPA/src/app/pages/Shop/com-documento/models/com-documento.ts
export interface ComDocumento {
	CORR_EMPRESA: number;
	ANIO_PERIODO: number;
	MES_PERIODO: number;
	CORR_TIPO_DOC: number;
	NOMBRE_TIPO_DOC: string;
	NOMBRE_CORTO_TIPO_DOC: string;
	CLASE_DOCUMENTO: string;
	SUMA_RESTA: number;
	CORR_DOCUMENTO: number;
  NUMERO_DOCUMENTO_CORR: string;
	NUMERO_DOCUMENTO: string;
	FECHA_DOCUMENTO: Date;
	FECHA_VENCIMIENTO: Date;
	CORR_TIENDA: number;
	CORR_PROVEEDOR: number;
	NOMBRE_PROVEEDOR: string;
	DESCRIPCION_PARTIDA: string;
	CORR_CONDICION_PAGO: number;
	NOMBRE_CONDICION_PAGO: string;
	DIAS_CREDITO: number;
	ESTADO_DOCUMENTO: string;
	NOMBRE_ESTADO_DOCUMENTO: string;
	ESTA_CONTABILIZADO: boolean;
	TOTAL_DOCUMENTO: number;
	TOTAL_NETO: number;
	SALDO_DOCUMENTO: number;
	CORR_TIPO_GASTO: number;
	NOMBRE_TIPO_GASTO: string;
	CANTIDAD: number;
	USUARIO_CREA: string;
	FECHA_CREA: Date;
	ESTACION_CREA: string;
	USUARIO_ACTU: string;
	FECHA_ACTU: Date;
	ESTACION_ACTU: string;
	CORR_MOVIMIENTO: number;
	CORR_MONEDA: number;
	FACTOR_CAMBIO: number;
	OPERADOR: string;
	SERIE: string;
	NUMERO_UNICO: number;
	ESTADO_ADMINISTRATIVO: string;
	NOMBRE_ESTADO_ADMINISTRATIVO: string;
	CODIGO_GENERACION: string;
	NUMERO_CONTROL: string;
	SELLO_RECEPCION: string;
  ANIO_PERIODO_IVA: number;
  MES_PERIODO_IVA: number;
}